Basic stat and StatInfluencedBy
Added basic stat (Intelligence, strength and dexterity)
Added a StatInfluenced attribut and isInfluencdeBy in stat so a particular stat can be influenced by an other component of the player (like level, dexterity, maxlife...)
Remove all the get function for attributs they are now embeded in the get part of their public getter

+++ Assets/Project/Scripts/Models/Abilities/Spell/LightningBall.cs
@@ -42,7 +42,7 @@ public class LightningBall : Ability
         ProjectileAbilityAttributs projectileAbilityAttributs = (ProjectileAbilityAttributs)abilityAttributs;
 
         // get the buffed value of projectile Number
-        int numberOfProjectile = (int)owner.getBuffedValue(projectileAbilityAttributs.numberOfBaseProjectile, StatType.NumberOfProjectile);
+        int numberOfProjectile = owner.getBuffedValue(projectileAbilityAttributs.numberOfBaseProjectile, StatType.NumberOfProjectile);
 
         // insert at the start of the description the number of projectile
         string lightningBallDescription = abilityAttributs.description.Replace("{0}", numberOfProjectile.ToString()) + "\n";
